@@ -79,46 +79,47 @@
   },
   "dependencies": {
     "@azure-rest/core-client": "^2.1.0",
-    "@azure/abort-controller": "^1.0.0",
-    "@azure/core-auth": "^1.7.2",
-    "@azure/core-lro": "^2.6.0",
-    "@azure/core-rest-pipeline": "^1.14.0",
+    "@azure/abort-controller": "^2.1.2",
+    "@azure/core-auth": "^1.9.0",
+    "@azure/core-lro": "^2.7.2",
+    "@azure/core-rest-pipeline": "^1.18.2",
     "@azure/core-tracing": "^1.2.0",
-    "@azure/logger": "^1.0.0",
-    "tslib": "^2.6.2"
+    "@azure/logger": "^1.1.4",
+    "tslib": "^2.8.1"
   },
   "devDependencies": {
     "@azure-tools/test-credential": "^2.0.0",
-    "@azure-tools/test-recorder": "^4.0.0",
+    "@azure-tools/test-recorder": "^4.1.0",
+    "@azure-tools/test-utils-vitest": "^1.0.0",
     "@azure/core-sse": "^2.1.3",
-    "@azure/core-util": "^1.9.0",
+    "@azure/core-util": "^1.11.0",
     "@azure/dev-tool": "^1.0.0",
     "@azure/eslint-plugin-azure-sdk": "^3.0.0",
-    "@azure/identity": "^4.3.0",
-    "@azure/monitor-opentelemetry-exporter": "1.0.0-beta.27",
+    "@azure/identity": "^4.6.0",
+    "@azure/monitor-opentelemetry-exporter": "1.0.0-beta.29",
     "@azure/opentelemetry-instrumentation-azure-sdk": "1.0.0-beta.7",
     "@opentelemetry/api": "^1.9.0",
     "@opentelemetry/instrumentation": "^0.57.1",
     "@opentelemetry/sdk-trace-node": "^1.30.1",
     "@types/node": "^18.0.0",
-    "@vitest/browser": "^2.0.5",
-    "@vitest/coverage-istanbul": "^2.0.5",
+    "@vitest/browser": "^3.0.9",
+    "@vitest/coverage-istanbul": "^3.0.9",
     "autorest": "latest",
     "dotenv": "^16.4.5",
     "eslint": "^9.9.0",
-    "playwright": "^1.41.2",
-    "source-map-support": "^0.5.9",
-    "typescript": "~5.7.2",
-    "vitest": "^2.0.5"
+    "playwright": "^1.50.1",
+    "typescript": "~5.8.2",
+    "vitest": "^3.0.9"
   },
   "scripts": {
-    "build": "npm run clean && dev-tool run build-package && dev-tool run vendored mkdirp ./review && dev-tool run extract-api",
+    "build": "npm run clean && dev-tool run build-package && dev-tool run extract-api",
     "build:samples": "echo Obsolete",
     "build:test": "npm run clean && dev-tool run build-package && dev-tool run build-test",
     "check-format": "echo skipped",
     "clean": "dev-tool run vendored rimraf --glob dist dist-* test-dist temp types *.tgz *.log",
+    "customize": "dev-tool customization apply -s ./generated -c ./src && npm run format",
     "execute:samples": "dev-tool samples run samples-dev",
-    "extract-api": "dev-tool run vendored rimraf review && dev-tool run vendored mkdirp ./review && dev-tool run extract-api",
+    "extract-api": "dev-tool run vendored rimraf review && dev-tool run extract-api",
     "format": "dev-tool run vendored prettier --write --config ../../../.prettierrc.json --ignore-path ../../../.prettierignore \"src/**/*.ts\" \"test/**/*.ts\" \"samples-dev/**/*.ts\" \"*.{js,json}\"",
     "generate": "dev-tool run vendored rimraf ./src/generated && dev-tool run vendored mkdirp ./src/generated && cp tsp-location.yaml ./src/generated && tsp-client update -o ./src/generated && dev-tool run vendored rimraf ./src/generated/tsp-location.yaml",
     "integration-test": "npm run integration-test:node && npm run integration-test:browser",
