@@ -3,11 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
+  "type": "module",
   "dependencies": {
     "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
@@ -28,8 +29,9 @@
     "jsonwebtoken": "^9.0.2",
     "jwt-decode": "^4.0.0",
     "lucide-react": "^0.477.0",
-    "next": "^14.1.0",
+    "next": "^15.2.0",
     "next-auth": "^4.24.11",
+    "next-intl": "^4.0.2",
     "next-seo": "^6.6.0",
     "nodemailer": "^6.10.0",
     "react": "^18.2.0",
@@ -46,7 +48,9 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^2.1.4",
+    "@next/eslint-plugin-next": "^15.2.4",
     "@types/jsonwebtoken": "^9.0.9",
+    "@types/negotiator": "^0.6.3",
     "@types/node": "^20",
     "@types/nodemailer": "^6.4.17",
     "@types/react": "^19",
