@@ -1,16 +1,16 @@
-import mysql from 'mysql2/promise';
+import mongoose from 'mongoose';
 import dotenv from 'dotenv';
 
 dotenv.config();
 
-const pool = mysql.createPool({
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
-  waitForConnections: true,
-  connectionLimit: 10,
-  queueLimit: 0
-});
+const connectDB = async () => {
+  try {
+    await mongoose.connect(process.env.MONGO_URI as string);
+    console.log('Kết nối MongoDB thành công');
+  } catch (error) {
+    console.error('Lỗi kết nối MongoDB:', error);
+    process.exit(1);
+  }
+};
 
-export default pool;
+export default connectDB;
